# Optimized DefaultObjectPool