@gfoidl 2018-02-20
> I also plan to create benchmarks projects for the primitives and this library. It would be great to also have the benchmarking code.

The code for this benchmarks is [here](https://github.com/gfoidl/Benchmarks/tree/master/aspnet/Common/ObjectPool). You can copy & paste it, or I could submit a PR when the benchmark project is created.